Feat: Automate random user choices in prompt
It's a chore to type in "rock", "paper", or "scissors" in a prompt
all the time.

So the `default` parameter of the `prompt(message, default)` function
makes it easy to get a default value.

The default is `${computerRandom}`, which is the computer's choice,
fetched from the function `computerPlay()`.

Other Changes:
- Update variable names to be more readable and meaningful (WIP).
  userChoiceValue variable is assigned to the return of promptMessage()
- const computerRandomValue = computerPlay();